requirements.txt
sentence-transformers>=3.0.0
rank-bm25>=0.2.2
faiss-cpu>=1.8.0
pandas>=2.2.0
numpy>=1.26.0
pyarrow>=15.0.0
tqdm>=4.66.0
pyyaml>=6.0.0

langgraph>=0.2.0
langchain-core>=0.3.0
langchain-ollama>=0.2.0
ollama>=0.3.0

fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sse-starlette>=2.1.0
pydantic>=2.8.0

streamlit>=1.38.0
httpx>=0.27.0

python-dotenv>=1.0.0
groq>=0.11.0
google-generativeai>=0.8.0

pytest>=8.0.0
